Time: 1 ms (99.98%) | Memory: 42.7 MB (46.23%) - LeetSync

+++ 7-reverse-integer/reverse-integer.java
@@ -1,7 +1,7 @@
 class Solution {
     public int reverse(int x) {
         long val=0;
-        long num=(long)x;
+  int num=x;
         while(num!=0){
             int temp=(int)num%10;
             val=val*10+temp;
